* Minor compile fixes for Windows and Windows mobile

+++ src/hagglekernel/ForwarderProphet.cpp
@@ -104,7 +104,7 @@ prophet_metric_t& ForwarderProphet::age_metric(prophet_metric_t& metric, bool fo
 		return metric;
 	}
 		
-	int K = (now - metric.second).getSeconds() / PROPHET_AGING_TIME_UNIT;
+	long K = (long)((now - metric.second).getSeconds() / PROPHET_AGING_TIME_UNIT);
 	
 	if (K > 0) {
 		double &P_ab = metric.first;
